website-configurations/data-parsing/requirements.txt
backports.zoneinfo==0.2.1
beautifulsoup4==4.13.3
bs4==0.0.2
Js2Py==0.74
numpy==1.24.4
pandas==2.0.3
psycopg2-binary==2.9.10
pyjsparser==2.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
six==1.17.0
soupsieve==2.6
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.2